Cleanup, move methods to more rational places
