+++ reddit_liveupdate/utils.py
@@ -4,10 +4,8 @@ import itertools
 import pytz
 
 from babel.dates import format_time, format_datetime
-from pylons import c, g
+from pylons import c
 from r2.lib import websockets
-from r2.lib.utils import extract_urls_from_markdown, domain
-from reddit_liveupdate.scraper import LiveScraper
 
 
 def pairwise(iterable):
@@ -49,34 +47,3 @@ def send_event_broadcast(event_id, type, payload):
     websockets.send_broadcast(namespace="/live/" + event_id,
                               type=type,
                               payload=payload)
-
-
-def embeddable_urls(md):
-    """ Given some markdown, return a list of all URLs that are from
-        liveupdate-embeddable domains.
-    """
-    return [u for u in extract_urls_from_markdown(md)
-            if domain(u) in g.liveupdate_embeddable_domains]
-
-
-def generate_media_objects(urls, maxwidth=485, max_embeds=15):
-    """ Given a list of embed URLs, scrape and return their media objects. """
-    media_objects = []
-    for url in urls:
-        scraper = LiveScraper.for_url(url)
-        scraper.maxwidth = maxwidth
-
-        # TODO: Is there a situation in which we would need the secure media
-        # object? Are twitter/youtube/imgur appropriately protocol agnostic?
-        thumbnail, media_object, secure_media_object = scraper.scrape()
-        if media_object and 'oembed' in media_object:
-            # Use our exact passed URL to ensure matching in markdown.
-            # Some scrapers will canonicalize a URL to something we
-            # haven't seen yet.
-            media_object['oembed']['url'] = url
-            media_objects.append(media_object)
-
-        if len(media_objects) > max_embeds:
-            break
-
-    return media_objects
